followup fix to timeout code - logging and debugging in test
